+++ demisto_sdk/commands/common/content/objects/pack_objects/playbook/playbook.py
@@ -1,7 +1,7 @@
 from typing import Union
 
 import demisto_client
-from demisto_sdk.commands.common.constants import PLAYBOOK
+from demisto_sdk.commands.common.constants import PLAYBOOK, FileType
 from demisto_sdk.commands.common.content.objects.pack_objects.abstract_pack_objects.yaml_content_object import \
     YAMLContentObject
 from wcmatch.pathlib import Path
@@ -21,3 +21,6 @@ class Playbook(YAMLContentObject):
             The result of the upload command from demisto_client
         """
         return client.import_playbook(file=self.path)
+
+    def type(self):
+        return FileType.PLAYBOOK