+++ demisto_sdk/commands/doc_reviewer/tests/doc_reviewer_test.py
@@ -1,0 +1,30 @@
+from typing import List
+
+from demisto_sdk.commands.doc_reviewer.doc_reviewer import DocReviewer
+
+
+def test_camel_case_split():
+    """
+    Given
+    - A CamelCase word
+    When
+    - Running camel_case_split on it.
+
+    Then
+    - Ensure result is a list of the split words in the camel case.
+    """
+    camel_1 = 'ThisIsCamelCase'
+    result = DocReviewer.camel_case_split(camel_1)
+    assert isinstance(result, List)
+    assert 'This' in result
+    assert 'Is' in result
+    assert 'Camel' in result
+    assert 'Case' in result
+
+    camel_2 = 'thisIPIsAlsoCamel'
+    result = DocReviewer.camel_case_split(camel_2)
+    assert 'this' in result
+    assert 'IP' in result
+    assert 'Is' in result
+    assert 'Also' in result
+    assert 'Camel' in result